Updates to TestHarness, Logger and ChildTest.
Addition of the changes I made to the TestHarness and the ChildTest. The Logger is not set up for using the message queue but has a reference to it for now. Need to update the Message struct to be able to handle the fields we want for getting Logger to the GUI.