Logger Changes
Adding functionality to the logger, including log level adding to the message and handling enqueueing the message to be sent to the GUI. Reworked some of the sprintf's as well as added some prevention for overflowing the MessageShort message field's buffer in the logger. Updated comments to be all one format. Added duration in ms to be added to the end of the test complete message.